packaged-program/requirements.txt
requests == 2.26.0
coverage
pytest-cov
mock
click